// intranet/src/app/core/helios-api-client/model/dataPagerOfMembreDTO.ts
/**
 * Helios API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MembreDTO } from './membreDTO';


export interface DataPagerOfMembreDTO { 
    total?: number;
    totalpage?: number;
    page?: number;
    size?: number;
    data: Array<MembreDTO>;
}