// intranet/src/app/core/helios-api-client/model/membreFiltre.ts
/**
 * Helios API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface MembreFiltre { 
    nom?: string | null;
    prenom?: string | null;
    email?: string | null;
    ville?: string | null;
    pays?: string | null;
    centre?: string | null;
    aspect?: string | null;
    statut?: string | null;
    l_villes?: Array<string> | null;
    l_pays?: Array<string> | null;
    l_centres?: Array<string> | null;
    l_aspects?: Array<number> | null;
    l_statuts?: Array<number> | null;
}