// intranet/src/app/core/helios-api-client/model/typeMembreDTO.ts
/**
 * Helios API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { NullableOfTypesMembres } from './nullableOfTypesMembres';


export interface TypeMembreDTO { 
    id: number | null;
    libelle: string | null;
    code: NullableOfTypesMembres | null;
}
export namespace TypeMembreDTO {
}